refactor(mapping): dedupe mapping-system harnesses (Rust + Python)
The mapping-system contract surface is unchanged; this is internal dedupe
across two pairs of files that had been copy-pasted.

Rust — extract shared harness into px-core::test_support
- New `test-support` feature on px-core gates a single home for the
  transform vocabulary (`direct`, `fixed_point_dollars`, `fixed_point_count`,
  `string_to_f64`, `parse_datetime`, `enum_remap`, `first_non_null`) and the
  walk-mapping-and-verify loop. Kalshi and Polymarket can no longer drift on
  what any transform means.
- Per-exchange `tests/mapping_contract.rs` shrink from 247/232 lines to 56
  each — they now own only the per-exchange mock setup; the verification
  hands off to `assert_mapping_contract`.
- Each exchange picks up `px-core` as a dev-dep with the `test-support`
  feature enabled; `serde_yaml` direct dev-dep dropped (now in px-core).

Python — extract shared mapping_lib.py
- `tools/mapping_lib.py` (55 lines) holds the file loaders and the JSON-
  pointer `resolve_ref` that `validate_mappings.py` and `render_mappings.py`
  had each copy-pasted.
- Both tools now import from it; semantic surfaces (type compatibility for
  validation, type-label rendering for docs) stay in their respective tools.

Net: ~−139 lines, but the structural win is bigger — the transform vocab
and the ref-resolution semantics each live in exactly one place.

- cargo test -p px-exchange-kalshi --test mapping_contract: ok
- cargo test -p px-exchange-polymarket --test mapping_contract: ok
- cargo test --workspace: ok (exit 0)
- cargo fmt --all --check: clean
- cargo clippy --workspace --all-targets: clean
- python tools/validate_mappings.py: PASS (same coverage)
- python tools/render_mappings.py: same 9,538 byte output
